Move RoutingService back to root of routing package
