+++ src/ext/java/org/opentripplanner/ext/transmodelapi/mapping/TransmodelMappingUtil.java
@@ -10,7 +10,7 @@ import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.MultiModalStation;
 import org.opentripplanner.model.Station;
 import org.opentripplanner.model.calendar.ServiceDate;
-import org.opentripplanner.routing.api.request.RoutingService;
+import org.opentripplanner.routing.RoutingService;
 
 import java.time.LocalDate;
 import java.util.Collection;
