+++ src/test/java/org/opentripplanner/updater/GtfsRealtimeFuzzyTripMatcherTest.java
@@ -2,7 +2,7 @@ package org.opentripplanner.updater;
 
 import com.google.transit.realtime.GtfsRealtime.TripDescriptor;
 import org.opentripplanner.GtfsTest;
-import org.opentripplanner.routing.api.request.RoutingService;
+import org.opentripplanner.routing.RoutingService;
 
 public class GtfsRealtimeFuzzyTripMatcherTest extends GtfsTest {
